fixed bug in precision for two real variables

+++ src/parameters_2d.f90
@@ -14,8 +14,8 @@ MODULE parameters_2d
   !> working precision
   INTEGER, PARAMETER :: wp = dp
 
-  REAL, PARAMETER :: four_thirds = 1.0_wp / 3.0_wp
-  REAL, PARAMETER :: neg_four_thirds = -1.0_wp / 3.0_wp
+  REAL(wp), PARAMETER :: four_thirds = 1.0_wp / 3.0_wp
+  REAL(wp), PARAMETER :: neg_four_thirds = -1.0_wp / 3.0_wp
 
   REAL(wp), PARAMETER :: tolh = 10.0_wp * EPSILON(1.0_wp)
